revert uneeded change for 32bit

+++ src/Table.cpp
@@ -59,7 +59,7 @@ static inline unsigned int Hash(void* key)
     // completely changed for 64-bit keys.
 #pragma warning( push )
 #pragma warning( disable : 4311 )
-    unsigned int a = (unsigned int)reinterpret_cast<unsigned long>(key);
+    unsigned int a = reinterpret_cast<unsigned int>(key);
 #pragma warning( pop ) 
     a = (a+0x7ed55d16) + (a<<12);
     a = (a^0xc761c23c) ^ (a>>19);
